refactor(apps/compatibility): moved app compatibility to models
