+++ open_wearable/lib/apps/models/app_compatibility.dart
